webgl: simplify and refactor WebGLContext
