--- lib/backends/webgl/webgl1-context.ts
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-
-import {BaseWebGLContext} from './base-webgl-context';
-
-export class WebGL1Context extends BaseWebGLContext {
-  constructor(
-      public canvas: HTMLCanvasElement, public gl: WebGLRenderingContext,
-      public contextAttributes: WebGLContextAttributes) {
-    super();
-    this.init();
-  }
-}
